# make the json log writer much faster