`go test -v -benchtime 10s -test.bench BenchmarkJSONFileLogger -test.benchmem -test.run=BenchmarkJSONFileLogger`

Before:
`BenchmarkJSONFileLogger-8	  200000	     66593 ns/op	  54.51 MB/s	     960 B/op	      30 allocs/op`
After:
`BenchmarkJSONFileLogger-8	  300000	     56547 ns/op	  64.19 MB/s	     960 B/op	      30 allocs/op`
Signed-off-by: Shijiang Wei <mountkin@gmail.com>