@GordonTheTurtle 2016-02-23
Job: Docker-PRs [FAILED](https://jenkins.dockerproject.org/job/Docker-PRs/24078/console):

~~~console

---
nfig 	python-dev 	python-mock 	python-pip 	python-websocket 	s3cmd=1.5.0* 	ubuntu-zfs 	xfsprogs 	libzfs-dev 	tar 	--no-install-recommends 	&& ln -snf /usr/bin/clang-3.8 /usr/local/bin/clang 	&& ln -snf /usr/bin/clang++-3.8 /usr/local/bin/clang++' returned a non-zero code: 100
Unable to find image 'docker:bf79b53' locally
Pulling repository docker.io/library/docker
Tag bf79b53 not found in repository docker.io/library/docker
Build step 'Execute shell' marked build as failure
Archiving artifacts
[PostBuildScript] - Execution post build scripts.
[Docker-PRs] $ /bin/sh -xe /tmp/hudson5405436810169475813.sh
+ set +x


---

~~~