@GordonTheTurtle 2016-02-23
Job: Docker-PRs-userns [FAILED](https://jenkins.dockerproject.org/job/Docker-PRs-userns/6562/console):

~~~console

---
o:490: DockerSuite.TestEventsRename	0.543s
PASS: docker_cli_events_test.go:439: DockerSuite.TestEventsResize	0.619s
PASS: docker_cli_events_unix_test.go:222: DockerSuite.TestEventsStreaming	0.509s
PASS: docker_cli_events_test.go:19: DockerSuite.TestEventsTimestampFormats	2.199s
PASS: docker_cli_events_test.go:500: DockerSuite.TestEventsTop	0.698s
PASS: docker_cli_events_test.go:46: DockerSuite.TestEventsUntag	2.659s

----------------------------------------------------------------------
FAIL: docker_cli_exec_test.go:22: DockerSuite.TestExec

docker_cli_exec_test.go:28:
    c.Assert(out, checker.Equals, "test")
... obtained string = ""
... expected string = "test"


----------------------------------------------------------------------
FAIL: docker_cli_exec_test.go:68: DockerSuite.TestExecAfterContainerRestart

docker_cli_exec_test.go:78:
    c.Assert(outStr, checker.Equals, "hello")
... obtained string = ""
... expected string = "hello"


----------------------------------------------------------------------
PASS:
---

~~~