@cpuguy83 2016-02-23
Could you add a 2nd benchmark with one reader subscribed while it's doing writes?