@calavera 2016-02-23
having those benchmarks directly in our tests would be great.