@mountkin 2016-02-27
The benchmark with reader is already added by @cpuguy83 https://github.com/docker/docker/blob/master/daemon/logger/jsonfilelog/jsonfilelog_test.go#L203